Code Cleanup part 3. (Read Desc).
Alr so I have done tons of little code improvements and I actually bothered to properly fix testnametags should work fine now :thumbsup: now I haven't tested but it SHOULD work perfectly now.

+++ src/main/java/me/earth/phobos/features/modules/combat/AntiCrystal.java
@@ -58,7 +58,7 @@ class AntiCrystal
         float highestDamage = 0.0f;
         for (Entity crystal : AntiCrystal.mc.world.loadedEntityList) {
             float damage;
-            if ( ! ( crystal instanceof EntityEnderCrystal ) || AntiCrystal.mc.player.getDistanceSq ( crystal ) > 169.0 || ( damage = DamageUtil.calculateDamage ( crystal , AntiCrystal.mc.player ) ) < this.minDmg.getValue ( ).floatValue ( ) )
+            if ( ! ( crystal instanceof EntityEnderCrystal ) || AntiCrystal.mc.player.getDistanceSq ( crystal ) > 169.0 || ( damage = DamageUtil.calculateDamage ( crystal , AntiCrystal.mc.player ) ) < this.minDmg.getValue ( ) )
                 continue;
             if ( bestcrystal == null ) {
                 bestcrystal = crystal;
